requirements.txt
Flask
requests
discord.py
